Make begin/end return Ruby version-aware offsets
As strings in Ruby prior to 1.9 reported their size in bytes rather than
characters, switch the specified behaviour of MatchData#begin and
MatchData#end to work regardless whether the offsets are bytes or
characters. This then allows us to switch them to return the appropriate
offset for the specific version of Ruby.

In order to calculate offsets efficiently, use rb_str_sublen where
possible to convert a byte offset to a char offset. However, while
rb_str_sublen exists in most Ruby > 1.8 implementations, it is
missing from Rubinius' C API. Add a configuration check to detect the
presence of the function and fallback to using rb_str_length on a new
string when it can't be used on an encoding-aware Ruby version.