Build against multiple re2 versions
To ensure backwards-compatibility, run the build against more versions
of Ruby but also two different versions of re2: an old one that does not
require C++11 and a recent one that does.

As we build against Ubuntu 12.04, we need to use clang to compile C++11
so override CC and CXX for those builds.