Fix infinite scan with empty patterns
GitHub: https://github.com/mudge/re2/issues/32

As FindAndConsumeN can return true without advancing through the given
input when matching against an empty pattern or a pattern with empty
capturing groups, we have to explicitly check when a match is made
against an empty string and advance ourselves.

This checks for both empty patterns and patterns with capturing groups
that match empty strings.

Adds a new eof? method to the Scanner to determine whether a scanner has
exhausted the input or not.