Use environment CC and CXX throughout compilation
GitHub: https://github.com/mudge/re2/issues/33

When compiling the gem with a compiler specified through the CC and CXX
environment variables, we need to update both the Makefile configuration
and standard Ruby configuration so that the various pre-compilation
checks and final Makefile use the right compiler toolchain.

This should hopefully allow compilation on OpenBSD using egcc and eg++
(tested in VirtualBox).