Stop using the deprecated utf8 API on RE2::Options
GitHub: https://github.com/mudge/re2/issues/40

As re2 has deprecated [0] and now removed [1] the utf8 option,
re-implement the option in the gem in terms of the encoding and
set_encoding API.

This should be entirely backward-compatible as the encoding API has been
present since the initial release in 2010.

Thanks to @buzzdeee for reporting this upcoming breaking change.

  [0]: https://github.com/google/re2/commit/58141dc9c92189ed8d046f494f5e034d5db91bea#diff-6dc69df7618951357bb5fb674c66aa34R640
  [1]: https://github.com/google/re2/commit/ac65d4531798ffc9bf807d1f7c09efb0eec70480#diff-6dc69df7618951357bb5fb674c66aa34L640-L649