Guard any other use of capturing groups
Following from the fix for RE2::Scanner#scan with an invalid regular
expression, ensure we always check a pattern is OK before using its
number of capturing groups as a positive integer.

This also fixes an edge case in RE2::Regexp#match when a user passes a
negative number for the number of intended matches. Previously, this
would throw a misleading NoMemoryError, now it throws a more informative
ArgumentError.