Silence STDERR when running specs
@pritambaral originally reopened STDERR to the platform-specific null
device but this was replaced in review by following how RSpec 3 captures
STDERR assigning Ruby's $stderr to a StringIO instance. However, as
suspected, this doesn't affect the underlying re2 library writing to
STDERR so its output would continue to print to the console.

The code for detecting the platform-specific null device is taken from
both the Ruby source for File::NULL [1] and the backports gem
implementation [2].

  [1]: https://github.com/ruby/ruby/blob/2d1cf658eeafbd989e0b61bae08a8d0b1b49e2bd/file.c#L6550-L6560
  [2]: https://github.com/marcandre/backports/blob/c21a80961ad162fcdd4b11ce56cf7c465af0f270/lib/backports/1.9.3/file/null.rb