Add :exception option to RE2::Set#match
In order to handle the different behaviour of RE2::Set::Match between
re2 ABI version 0 and all later versions, add an :exception option to
RE2::Set#match that is true by default. This will raise detailed errors
as provided by RE2::Set::Match or raise an RE2::Set::UnsupportedError if
the ABI version of re2 does not support it.

Setting :exception to false will no longer raise errors and is therefore
available to all ABI versions of re2.

This way, we are somewhat close to the underlying re2 API (where
RE2::Set::Match can optionally populate ErrorInfo) and warn callers as
soon as possible if exceptions cannot be raised with their current
version of the re2 library.