Support static linking of re2 library
Currently when the re2 system library is updated by Homebrew, the
dynamically linked library is removed and moved to a new location.
This breaks the re2 gem and requires users to run `gem pristine re2`
to get back into a working state.

To avoid this hassle, this commit adds a `--enable-static` flag that
will statically link the library in the gem's C extension library.

Homebrew doesn't currently install a static library for re2. This will
be fixed via https://github.com/Homebrew/homebrew-core/pull/119928.