Make cross-compilation for native gems work
mini_portile2 provides `configure` script with the `--host` parameter,
and the script figures out the target compilers to use. With CMake,
we don't have that benefit so we need to set the C and C++ compilers
ourselves.

We use a simple heuristic to find the target compilers:

1. For macOS, we use `clang` and `clang++. Otherwise we use `gcc` and
   `gcc++`.
2. We search the PATH for `<host>-compiler`. Use that if we find it.
   Otherwise default to the base compilers.

To make CMake work with cross-compilation, a number of variables have
to be set:

- CMAKE_SYSTEM_PROCESSOR
- CMAKE_SYSTEM_NAME
- CMAKE_C_COMPILER
- CMAKE_CXX_COMPILER