Fix static linking when system libraries installed
Ruby's mkmf appears to list `RbConfig::CONFIG['exec_prefix']/lib`
first in the search path when building a C extension. If you have
libre2 installed in that path, the linker will use that library
instead of the vendored library. This causes an undefined symbol error
when using the extension.

```
/usr/local/bin/ruby: symbol lookup error: /usr/local/bundle/gems/re2-2.0.0.beta1/lib/re2.so: undefined symbol: _ZN3re23RE2C1ESt17basic_string_viewIcSt11char_traitsIcEERKNS0_7OptionsE
```

This commit does two things to ensure the final shared library is
compiled statically with the vendored libraries:

1. For -L<path> flags, ensure that any `ports` paths are prioritized just
   in case there are installed libraries that might take precedence.

2. For -l<lib> flags, convert the library to the static library with a
   full path and substitute the absolute static library.  For example,
  -lre2 maps to /path/to/ports/<arch>/libre2/<version>/lib/libre2.a.