Add Ruby 2.6 support for native gems
As the default system Ruby on macOS Monterey is 2.6 and MiniPortile2
supports back to 2.3, restore support for that Ruby and compile the C
extension for it in the native gems.

The trade-off here is that we increase the size of the native gem in
order to support more Ruby versions. While 2.6 is technically EOL, the
rubygems.org stats still show usage of it in the past month:
https://ui.honeycomb.io/ruby-together/datasets/rubygems.org/result/HBSqTboW1yi

Note Ruby 2.6 triggers a register error with C++17 (it was fixed by
https://github.com/ruby/ruby/commit/113bef697633803a33a547914b16ba5fbef165b8
in Ruby 2.7) so we disable that particular error during compilation.