Remove unused Rake task to override version
To simplify the build pipeline, building gems will use whatever version
is specified in the gemspec (which, in turn, comes from
lib/re2/version.rb). Rather than dynamically changing this for certain
CI jobs, remove the task altogether even though this increases the risk
of downloading and publishing the "wrong" gem for maintainers.