Restore Ruby 2.5 support
While trying to justify the restoration of Ruby 2.6 support by looking
at the stats from RubyGems.org, it appears that the third most popular
Ruby version is still 2.5. It's not clear why this is but if it is easy
for us to add support, this means we cover the top five most used Ruby
versions according to
https://ui.honeycomb.io/ruby-together/datasets/rubygems.org/result/HBSqTboW1yi
(with the bonus of supporting 2.6 due to it being the default system
Ruby on macOS Monterey).