Use references to NamedCapturingGroups, not copies
As RE2's NamedCapturingGroups returns a const reference to its map of
names to indices, use that rather than immediately taking a copy. Switch
to using iterators over the data where possible. Note we don't use auto
for the types as that isn't available when compiling against the oldest
versions of RE2.