Use RSTRING_PTR when we know the arg is a String
As we're now calling StringValue earlier in functions, there's no need
to duplicate the coercion with StringValuePtr when we can use
RSTRING_PTR directly instead.

Also, as raising an exception will skip calling any C++ destructors, we
need to be more intentional with when we use macros like StringValue and
StringValuePtr as they may raise after allocating memory.