Automatically update docs site
Rather than manually updating the docs at https://mudge.name/re2,
configure GitHub Actions to do it automatically on every push to the
main branch.

Note we have to install redcarpet for YARD's Markdown support but
redcarpet doesn't compile on Alpine so we keep it and YARD out of the
gem's dependencies.