Add endpos argument to RE2::Regexp#match
GitHub: https://github.com/mudge/re2/issues/119

Expose RE2::Match()'s endpos argument in Ruby so users can specify an
offset at which to stop matching.

Note that old versions of RE2 don't accept an endpos argument when
matching so we explicitly detect this and raise an exception when
attempting to pass it to a version that doesn't support it.