Precompile native gems for Ruby 3.3.0-preview3
Use [rake-compiler-dock
1.4.0.rc1](https://github.com/rake-compiler/rake-compiler-dock/releases/tag/1.4.0.rc1)
to precompile a native gem for Ruby 3.3 and test it on all supported
platforms.

This won't be safe to release until Ruby 3.3's ABI is guaranteed not to
change (the recently released 3.3.0-rc1 shows the first sign of ABI
stability but that isn't supported by rake-compiler-dock yet).

Note we still can't test Ruby 3.3 on Windows 2022 as setup-ruby doesn't
support it yet and we need to use the full name of "3.3.0-preview3" for
Docker image tags.