Expand platform coverage in CI
To ensure we cover as many platforms as possible, add a dedicated
FreeBSD job to CI and ensure we test both --enable-system-libraries and
--disable-system-libraries where possible.

We upgrade to MiniPortile2 2.8.6 as this fixes the default compiler on
FreeBSD (otherwise we'd have to specify CC and CXX ourselves).

Note we can now remove our pkgconf version pin as 2.2.0 (which fixes the
dependency ordering) is now widely available.