Precompile for Ruby 3.4.0-rc1
Upgrade to rake-compiler-dock 1.7.0.rc1 and precompile the native gem
for Ruby 3.4.0-rc1 assuming this will be representative of Ruby 3.4.0
which will be released on Christmas Day.

Switch to use Ruby 3.4.0-rc1 as the default version of Ruby in CI but
use Ruby HEAD when testing on Windows as pre-release versions of Ruby
are not made available for that platform.