Restore Ruby 2.6, 2.7, and 3.0 support
Upgrade to rake-compiler-dock 1.8.0 which allows us to compile Musl gems
for Ruby 3.0 (see https://bugs.ruby-lang.org/issues/17788) and restore
support for Ruby 2.6 and 2.7 at the same time. This requires installing
the latest supported version of RubyGems for older versions of Ruby.

We now switch to Ruby 3.4 final (rather than using the release
candidate) on all platforms for both precompilation and testing.

This reverts commit a0ec2d6cfaa6115a02eccdb3ccd02ed1938e0136.